merge both junit mocks into one
Merging reduces code duplication and makes usage easier. The old junit5 class
is still kept for backwards-compatibility for now, but marked deprecated.
At some point, it will be completely removed.

Additionally, the explicit JWKS response validation is removed from the junit
module tests, because this is the responsibility of the mock module to verify.

Signed-off-by: Kai Helbig <kai.helbig@tngtech.com>