support cookie re-authentication
once a session is created and a session cookie is in place, any request to the
mock server will directly be redirected and re-use the existing session data

This is e.g. necessary to support cypress tests, e.g. when the test logs in
before loading the actual page, and loading the page causes a second login
attempt.
With this change, the second attempt is directly redirected back to the page,
because the login session is still valid.

For now, we do not support client requests for re-authentication (e.g. we ignore
redirects where acr=1 is requested, which in theory should require the user to
login again). Also, a session's scope is global, not bound to a single client.

Signed-off-by: Kai Helbig <kai.helbig@tngtech.com>