allow individual configuration of realm and hostname per token
issue #61

It is now possible to configure the issuer realm and host per generated token.
As a side effect, the server configuration parameters realm and hostname have
been renamed to defaultRealm and defaultHostname. The old methods are still
present but have become deprecated and will be removed in a sub-sequent release.

Signed-off-by: Kai Helbig <kai.helbig@tngtech.com>