split session handling to allow immutable objects
The initialization of a new session consists of two steps:
- opening the login page with the request parameters from the client
- logging in and either directly or via authorization code offering
  an access token

To make this process more clear in the code, the session repository
now distinguishes between an attempt to create a session (= session
request) that is created for any visitor of the login page, and an
actual session that is created once the user has provided his name
and desired roles and that is persisted for future use.

This change is also a preparation for the password grant flow, where
no session requests and no persistent sessions are created.

Signed-off-by: Kai Helbig <kai.helbig@tngtech.com>