change the way roles are assigned for login
Instead of defining a list of resources to apply the roles to, the user can now
define a list of default audiences to add. In a new, separate setting, the user
can decide whether the roles given by the login shall be applied to the realm,
all audiences (which double as resources), or both.

This is a breaking API and behavior change!

Signed-off-by: Kai Helbig <kai.helbig@tngtech.com>